backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
databases[postgresql]==0.8.0
asyncpg==0.29.0
pandas==2.1.4
numpy==1.24.4
yfinance==0.2.28
TA-Lib==0.4.28
APScheduler==3.10.4
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2